+++ core/mixer/image/image_mixer.cpp
@@ -235,7 +235,7 @@ private:
 		kernel_.draw(std::move(draw_params));
 	}
 			
-	safe_ptr<device_buffer> create_mixer_buffer(size_t stride, const video_format_desc& format_desc)
+	safe_ptr<device_buffer> create_mixer_buffer(uint32_t stride, const video_format_desc& format_desc)
 	{
 		auto buffer = ogl_->create_device_buffer(format_desc.width, format_desc.height, stride);
 		ogl_->clear(*buffer);
