+++ core/producer/channel/channel_producer.cpp
@@ -99,7 +99,7 @@ public:
 		return false;
 	}
 
-	virtual size_t buffer_depth() const override
+	virtual uint32_t buffer_depth() const override
 	{
 		return 1;
 	}
